requirements.txt
azure-storage-blob
youtube-dl